requirements.txt
fastapi
uvicorn[standard]
python-multipart
pymupdf
pdfplumber
pdf2image
Pillow
python-doctr
spacy
presidio-analyzer
presidio-anonymizer
usaddress
pyyaml
reportlab
openai
python-dotenv
requests
transformers
torch
